feat(boot): Added basic serial debugging functions
Added a basic kprintf function for printing text to the serial port via
the UART class. Also added an ASSERT preprocessor for comparing expected
and actual values in the kernel. This prints a message to the serial
port and exits in the case of a failure. This makes it much easier to
check basic issues during boot without having to run in debug mode.