fix: Fixed incorrect multiboot info struct pointer
Modified the linker.ld to fix issues caused after updating gcc version.
I don't know why or how this fixed the issues. It is clear that my
understanding of the linker script needs improving as it seems to be a
constant cause of issues that I don't really understand. :(

Temp fix to issues beginning in commit 287f574