feat(boot): Kprint bootloader name and cmd options
The kernel will now print the name of the bootloader and any command
line parameters passed to it from grub. This will be used in the future
for configuring early boot parameters such as whether or not to debug to
the serial port. The grub config file has been updated with some example
config parameters.