feat(std): Added strnlen function
Added a strnlen function to string.h. This is a variant of strlen with
an additional max length arguement. The function shall return the string
length just like strlen, however it will stop at the max length
specified and return that instead. This function is not part of the C
standard however is used in kprintf and has similar syntax to existing
string.h functions.