feat(trace): Added a basic stack trace debugger
Added a basic stack trace debugger that walks back through the stack
frames printing out their addresses. Also added a rule to the makefile
to generate a table of symbols. This isn't yet used by the stack trace
but it is still convenient even for manually tracing the stack. This
function will likely be moved, probably with the LOG defines.