refactor(printf): Added _pad() and removed goto
Added a _pad() function for padding the buffer or output with a defined
character. This allows the buffer overrun checks in _printf_impl() to
break rather than goto, as they are no longer inside of a while loop.
Additionally renamed some variables for clarity and added
documentation/comments to internal functions. Finally, removed the
int2buf() function and just embedded it in the main _print_impl()
function as it was only ever called from one place.