feat(cxxabi): Added __cxa_pure_virtual()
Added the __cxa_pure_virtual() function to the new cxxabi files. This is
supposed to enable the use of pure virtual functions according to the
OSDev wiki, but it seemed to work without it. Regardless, including this
should allow the kernel to produce an error message if a pure virtual
function is somehow called.