build: Started transitioning from make to cmake
Started transitioning the project from using make to cmake. Hopefully
this makes it easier to manage the project as it gets larger. At the
moment it compiles kernel.bin succesfully but does not produce and iso.
This also required disabling _init and _fini inside init64_start as I
haven't figured out how to link crtbegin.or and crtend.o, although it
seems the functions weren't doing anything anyway.