build: More work on transitioning to cmake
Made more progress towards to transitioning to cmake. This mostly
includes modifying the old make file to just produce iso images (to be
added to cmake), as well as including the crtbegin.o and crtend.o (these
aren't linked in the correct order but seem to work, or atleast not
crash).