feat(stdio): Added nullptr case to printf '%p'
Added a special case to printf to print "(nullptr)" when the arguement
has the pointer specifier and the value of the arguement is 0 (nullptr).
If the user wishes to print null pointers using the standard 0x...0
format, they can use "%#.16x" as the specifier instead. Note, that the
printf specifications says pointers are printed in "an
implementation-defined manner" thus, this still meets the spec.