misc: Updated VS code settings for gcc 13.2.1
Updated various VS code settings for gcc 13.2.1, this includes the file
associations with libstdc++ headers. Note, most of these headers are not
suitable for use in the kernel. e.g. std::optional.value() can throw
(which will not compile) whereas the custom optional triggers a kernel
panic instead. It may be worth renaming the custom header namespaces to,
for example, kstd:: for clarity.