refactor: Renamed namespace std to kstd
Renamed the std namespace in libc++ to kstd to distinguish it from the
std namespace supplied by the compiler in libstdc++. Additionally,
deleted the libc wrapper libraries now that libc++ is not strictly
following the standard. I am still undecided on whether or not to use
std::utility or kstd::utility.