feat(libc++): Added basic pair and tuple headers
Added some basic pair and tuple headers. Neither are implemented and
only contain the constructors. They also both depend on each other, so
some re-shuffling of declarations will likely be needed. Additionally,
paid should be defined in utility, but it might make sense to define it
as a standalone header as well.