feat(ksyms): Modified get_symbol to return a pair
Modified KSyms::get_symbol() to return a kstd::pair<const char *,
uintptr_t> instead of returning a const char * and passing a
uintptr_t* as an arguement. This makes the code more expressive and
easier to understand. Additionally Debug::trace_stack() was updated to
make use of the new syntax.