feat(libc++): Add rebind to kstd::allocator
Added a rebind member to the kstd::allocator class. This can be used to
bind the allocator to a different type. This was part of the C++
standard but was deprecated in favour of allocator_traits. For now we
will use the old standard. The syntax for rebinding is as below.

using NewAlloc = typename OldAlloc::template rebind<new_type>::other;