feat: Add write-combining PAT entry for framebuffer
Added temporary code to main for adding a write-combining entry to the
Page Attribute Table (PAT). This entry is then used by the framebuffer
when mapping the framebuffer pages into memory. This makes writing to
the screenbuffer significantly faster as the buffer is only ever written
to thus it does not require strong ordering.