feat(libc++): Add basic kstd::bitfield class
Started implementing a simple bitfield class. This class simply contains
a templated type, e.g. uint64_t, and provides functions for
accessing/modifying individual bits in the underlying type. This class
is not part of the C++ standard library (as far as I know), but should
still be useful within the kernel.