feat(memory): Add bitmap frame allocator
Replaces the temporary bump frame allocator with a more robust, less
temporary bitmap allocator. This allocator simply stores a bitmap for
each memory region that can expand over time as more frames are
allocated. This will likely not be the final form of the frame allocator
as it quite slow compared to, for example a buddy allocator, however
will be more than sufficient for the time being.