feat(memory): Avoid copying memory regions vector
Added a function to the Memory namespace for getting a reference to the
memory regions vector. This allows FrameAllocator to access the memory
regions vector without needing to copy it during initialization. The
getter function needed to be defined in a seperate header due to
circular dependencies with the vector class.