feat(boot): Start configuring kernel for high half
Started reconfiguring the kernel to be mapped into thee high half. So
far I have managed to get it to compile assuming it is in the high half,
and it manages to get back into long mode, however it can't yet jump to
the high address (currenlty tripples faults). The current assumption is
that the page tables for the high half are incorrect. Additionally, once
that is fixed, the paging system will need to be reworked as the high
half interferes with the recursive paging. Finally, in order to get it
to compile I had to not include the _init and _fini sections as libgcc
needs to be rebuilt with mcmodel=kernel variants.