fix: GDT pointer in higher half
Fixed an issue where the GDTR pointer was still pointing to the lower
half mapped kernel. Additionally, merged header.asm, init32.asm and
init64.asm into a single boot.asm as marking various gdt and stack
fields extern just for init64 to see them seems over the top. Finally,
identified an issue when trying to unmap the lower half kernel. This
causes a page-fault -> double-fault -> tripple-fault. It is still
unclear why this fault is occuring as the kernel was mapped to the
higher half.