fix: kstd::string works with wchar_t (and others)
Fixed to bugs within the kstd::string class preventing it from working
with char types larger than 1 byte. First, the calculation for the number
of chars in the small string buffer did not account for the size of the
char. Second, the class was making use of strlen internally which only
supports 1 byte char types. This is normally avoided by using the
char_traits class which works for any size char, however for simplicity,
a seperate internal __strlen function was defined instead.