feat(libc): Add wide char support to printf
Added wide char support to all of the printf functions. These functions
now fulfill the entire spec except for floats which aren't enabled in
the kernel. This will be added within an #ifndef __is_kernel sometime in
the future. Additionally, the spec says that precision should affect the
byte length of a string rather than the char length. In my opinion it
makes more sense to use the char length as formatting text would be
easier (atleast with monospace text). The size arguement used by
snprintf still respects byte length as this would produce a buffer
overflow otherwise. For now atleast, this implementation uses char
length for precision.