fix(libc): Correct mask in mbrtowc()
Fixed the mask used in mbrtowc() function within wchar.h. Previously the
mask used was just removing the leading 1's used to indicate the number
of bytes, now it also includes the 0 that follows as this is not part of
the data. This likely wouldn't have mattered as that bit should always
be 0 thus not affecting the value, but the fix makes the solution more
correct.