fix: Propogate errno not redeclare
Modified wcsnrtombs() and mbsnrtowcs() functions to simply propogate the
errno value set by the wcrtomb() and mbrtowc() functions respectively
rather than redeclare them. In this case there is no difference, but if
more advanced encodings that actually make use of the mbstate_t are
added in the future then the functions may also set errno to EINVAL
which would have been overwritten.