fix: Correct frame allocation when modules > 2 MiB
Fixed an issue where the frame allocator was allocator physical pages
that were actually already occupied by any modules loaded. This only
affected the kernel if the modules were greater than 2 MiB as this would
make them to large to fit in the space before the kernel. For now this
fix simply gets the modules entry from the multiboot info table, however
this should be changed to referencing some form of InitRD API instead.