feat(libc++): Minor improvements to string_view
Modified the constructor for basic_string_view such that when it is
given a nullptr but no size, it simply initializes the view as if it
were empty rather than fault. This is not the same behaviour as defined
in the C++ standard which simply states that this case is undefined
behaviour. Additionally added equality operators for c-style strings.
Again this is also not in the C++ standard.