refactor: Remove __attribute__((unused))
Removed the __attribute__((unused)) (or cast to void) for functions that
were not making use of an arguement. This was only done on certain
functions to silence the compiler warnings when the arguements
explicitly needed to be defined (due to library or hardware spec).
Instead, the variable should simply not be given a name as this also
silences the compiler warnings and is much shorter.