fix: Page fault and scheduler lock issues resolved
Fixed two issues introduced by the switch to std::list from std::vector.
The first issue was that if there was too much time between
Scheduler::init() and Scheduler::start(), the scheduler might be invoked
by the timer causing a page fault as the list might still be empty. This
was fixed by delaying the interrupt unmasking until the first task was
running. This means that their is now an idle task (ID 0) internal to
the scheduler that is responsible for unmasking the interrupt and then
continuously yielding the scheduler. In the future this task will also
be responsible for cleaning up stopped tasks.

The second issue was that the global current_task iterator was not
actually updated, causing the scheduler to always start searching for a
new task from the beginning of the list resulting in it always picking
the same task.