feat(scheduler): Add sleep_for() and sleep_until()
Added the functions Scheduler::sleep_until() and Scheduler::sleep_for().
These functions set the current thread state to Sleeping and append it
to a priority queue ordered by the wake up time. This queue is the
queried every time the scheduler is invoked. For now, the wake up time
is based on the CPU TSC and uses a hardcoded CPU frequency value to
convert the TSC into nanoseconds.