feat: Enable SSE instructions
Added code to Kernel::main() for enabling SSE instructions in the CPU.
This also allows us to now use floats as gcc uses the SSE registers by
default instead of the FPU for x86_64. Note however, that SSE registers
are not yet stored between context switches, as the address arguement
for fxsave/fxrstor must be 16-byte aligned, and alignment is not yet
supported by std::list (container of threads).