fix: Correct context template for format_string
Fixed an issue where the basic_format_string constructor was
incompattible with the __vformat_to() function. This is because
__vformat_to expects Iter to either be the same as __format_iter or can
be passed into __format_iter_buf (not constexpr). To fix this, the
context portion of the formatting function was extracted to __do_format
which is now used by the basic_format_string constructor instead.