feat(libc++): Implement default std::formatter
Implemented the default std::formatter. This formatter simply outputs
the name of the type and it's address. It currently only works with Char
= char as __PRETTY_FUNCTION__ is evaluated as a const char *. If it
tries to get the type with Char = wchar_t the type name will just be
returned as L"unknown"; Additionally, this formatter doesn't actually
print anything useful at the moment as it depends on the string and
pointer formatters.