fix: Make basic_format_arg constructor constexpr
Fixed an issue where the constructor of std::basic_format_arg was not
considered constexpr when using a non-trivial type (i.e. string_view or
basic_format_arg::handle). This was not previously considered an error
with GCC 13 but is now with GCC 14. The issue was fixed by initializing
the union values using std::construct_at().