feat: Use PIT ticks for scheduling
Modified the scheduler to use the PIT with a counter to keep track of
how many ticks have passed since boot. The interface for the tick
counter will likely be moved to a seperate Time module. Additionally,
the sleep_for() and sleep_until() functions that make use of nanoseconds
will be moved the std::thread.